+++ src/interface/IFeeManager.sol
@@ -12,10 +12,10 @@ interface IFeeManager is ILedgerComponent {
         FuturesFeeCollector
     }
 
-    function getOperatorGasFeeBalance(bytes32 tokenHash) external returns (uint256);
-    function setOperatorGasFeeBalance(bytes32 tokenHash, uint256 amount) external;
-    function getFeeAmount() external returns (uint256);
-    function changeFeeAmount(uint256 amount) external;
+    function getOperatorGasFeeBalance(bytes32 tokenHash) external returns (uint128);
+    function setOperatorGasFeeBalance(bytes32 tokenHash, uint128 amount) external;
+    function getFeeAmount() external returns (uint128);
+    function changeFeeAmount(uint128 amount) external;
     function getFeeCollector(FeeCollectorType feeCollectorType) external returns (bytes32);
     function changeFeeCollector(FeeCollectorType feeCollectorType, bytes32 _newCollector) external;
 }
